% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_rdoc_gg.R
\name{plot_rdoc}
\alias{plot_rdoc}
\title{Alias for \code{\link[=plot_rdoc_gg]{plot_rdoc_gg()}}}
\usage{
plot_rdoc(corr_df, ...)
}
\arguments{
\item{corr_df}{Data frame with columns \code{Domain}, \code{Term}, \code{r}, and optional \code{p}.
\code{Term} must be unique.}

\item{...}{Additional arguments passed to \code{\link[=plot_rdoc_gg]{plot_rdoc_gg()}}.}
}
\value{
A ggplot object.
}
\description{
Alias for \code{\link[=plot_rdoc_gg]{plot_rdoc_gg()}}
}
