% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette_utils.R
\name{rdoc_available_palettes}
\alias{rdoc_available_palettes}
\title{List Available Domain Palettes}
\usage{
rdoc_available_palettes()
}
\value{
Character vector with available palette names.
}
\description{
Returns palette names supported by \code{rdocodeR}.
}
\examples{
head(rdoc_available_palettes())
}
