% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc_data.R
\name{rdoc_example_data}
\alias{rdoc_example_data}
\title{Read Example RDoC Correlation Data}
\usage{
rdoc_example_data()
}
\value{
A data frame.
}
\description{
Loads the bundled TSV example data with columns \code{Domain}, \code{Term}, \code{r}, and \code{p}.
}
