% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdoc_data.R
\name{rdoc_terms_file}
\alias{rdoc_terms_file}
\title{Path to Bundled RDoC Term Maps}
\usage{
rdoc_terms_file()
}
\value{
Absolute file path to the bundled term map RDS file.
}
\description{
Returns the path of the bundled \code{rdoc_terms.fsavg6.lh.rh.rds} file.
}
